+++ Order.Management/Models/Toy.cs
@@ -1,15 +1,15 @@
 namespace Order.Management.Models
 {
-  public struct Toy
-  {
-
-    public ToyShape Shape;
-    public ToyColor Color;
-
-    public Toy(ToyShape shape, ToyColor color)
+    public struct Toy
     {
-      Shape = shape;
-      Color = color;
+
+        public ToyShape Shape;
+        public ToyColor Color;
+
+        public Toy(ToyShape shape, ToyColor color)
+        {
+            Shape = shape;
+            Color = color;
+        }
     }
-  }
 }
