+++ Order.Management/Models/ToyShape.cs
@@ -1,9 +1,9 @@
 namespace Order.Management.Models
 {
-  public enum ToyShape
-  {
-    Square,
-    Circle,
-    Triangle
-  }
+    public enum ToyShape
+    {
+        Square,
+        Circle,
+        Triangle
+    }
 }
